Remove TODOs (decided not to change for now)
Moving those methods to the normal state would complicate error
handling. What would happen if some error occurred within the
section where the actual module is in program mode? It may leave the
module in program mode, and with some transaction half-pending.
I prefer to have to switch to the mode and back, but this might change